requirements.txt
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
Django==4.2
django_debug_toolbar==3.8.1
Faker==12.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.3.0
mccabe==0.7.0
mixer==7.2.2
packaging==25.0
pep8-naming==0.13.3
pluggy==1.6.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.9.0.post0
pytz==2022.7
six==1.17.0
snowballstemmer==3.0.1
sqlparse==0.5.5
tomli==2.4.0
tzdata==2025.3
